refactor: split audit package for protocol independence
Move core types and publishing logic to pkg/audit, HTTP middleware to
pkg/audit/httpaudit. This prepares for future gRPC audit support without
locking the package hierarchy to HTTP.

- pkg/audit: types (Payload, Actor, HTTP*), options (WithAuth,
  WithOrganizationID, WithStackID), event publishing helpers
- pkg/audit/httpaudit: HTTP middleware, sensitive paths, response writer
